requirements.txt
openai
langchain
faiss-cpu
python-dotenv
streamlit
pypdf
tiktoken
chromadb
langchain-community
